+++ affinidi-messaging-mediator/src/database/handlers.rs
@@ -2,16 +2,15 @@ use super::DatabaseHandler;
 use crate::common::{config::Config, errors::MediatorError};
 use deadpool_redis::Connection;
 use redis::aio::PubSub;
-use std::{thread::sleep, time::Duration};
+use std::{fs::read_to_string, thread::sleep, time::Duration};
 use tracing::{error, event, info, Level};
 
-const LUA_SCRIPTS: &str = "atm-functions.lua";
 const REDIS_VERSION: &str = "7.4"; // Minimum Redis version required
 
 impl DatabaseHandler {
     pub async fn new(config: &Config) -> Result<Self, MediatorError> {
         // Creates initial pool Configuration from the redis database URL
-        let pool = deadpool_redis::Config::from_url(&config.database_url)
+        let pool = deadpool_redis::Config::from_url(&config.database.database_url)
             .builder()
             .map_err(|err| {
                 event!(Level::ERROR, "Database URL is invalid. Reason: {}", err);
@@ -25,11 +24,11 @@ impl DatabaseHandler {
         // and create the async pool of redis connections
         let pool = pool
             .runtime(deadpool_redis::Runtime::Tokio1)
-            .max_size(config.database_pool_size)
+            .max_size(config.database.database_pool_size)
             .timeouts(deadpool_redis::Timeouts {
-                wait: Some(Duration::from_secs(config.database_timeout.into())),
-                create: Some(Duration::from_secs(config.database_timeout.into())),
-                recycle: Some(Duration::from_secs(config.database_timeout.into())),
+                wait: Some(Duration::from_secs(config.database.database_timeout.into())),
+                create: Some(Duration::from_secs(config.database.database_timeout.into())),
+                recycle: Some(Duration::from_secs(config.database.database_timeout.into())),
             })
             .build()
             .map_err(|err| {
@@ -42,7 +41,7 @@ impl DatabaseHandler {
 
         let database = Self {
             pool,
-            redis_url: config.database_url.clone(),
+            redis_url: config.database.database_url.clone(),
         };
         loop {
             let mut conn = match database.get_async_connection().await {
@@ -84,32 +83,7 @@ impl DatabaseHandler {
         _check_server_version(&database).await?;
 
         // Check and load LUA scripts as required
-        {
-            let mut conn = database.get_async_connection().await?;
-            let function_load: Result<String, deadpool_redis::redis::RedisError> =
-                deadpool_redis::redis::cmd("FUNCTION")
-                    .arg("LOAD")
-                    .arg("REPLACE")
-                    .arg(LUA_SCRIPTS)
-                    .query_async(&mut conn)
-                    .await;
-            match function_load {
-                Ok(function_load) => {
-                    event!(
-                        Level::INFO,
-                        "database response for FUNCTION LOAD: ({})",
-                        function_load
-                    );
-                }
-                Err(err) => {
-                    event!(
-                        Level::WARN,
-                        "database response for FUNCTION LOAD: ({})",
-                        err
-                    );
-                }
-            }
-        }
+        _load_scripts(&database, &config.database.functions_file).await?;
 
         database.get_db_metadata().await?;
         Ok(database)
@@ -202,5 +176,51 @@ async fn _check_server_version(database: &DatabaseHandler) -> Result<String, Med
             "NA".into(),
             "Couldn't find redis_version in server info".into(),
         ))
+    }
+}
+
+// Private Helper function to load Redis scripts into the database
+async fn _load_scripts(
+    database: &DatabaseHandler,
+    scripts_path: &str,
+) -> Result<(), MediatorError> {
+    // Load the file contents into a string
+    let lua_scripts = read_to_string(scripts_path).map_err(|err| {
+        MediatorError::ConfigError(
+            "Initialization".into(),
+            format!(
+                "Couldn't ready database functions_file ({}). Reason: {}",
+                scripts_path, err
+            ),
+        )
+    })?;
+
+    let mut conn = database.get_async_connection().await?;
+    match deadpool_redis::redis::cmd("FUNCTION")
+        .arg("LOAD")
+        .arg("REPLACE")
+        .arg(lua_scripts)
+        .query_async::<String>(&mut conn)
+        .await
+    {
+        Ok(_) => {
+            event!(
+                Level::INFO,
+                "Loaded LUA scripts into the database from file: {}",
+                scripts_path
+            );
+            Ok(())
+        }
+        Err(err) => {
+            event!(
+                Level::WARN,
+                "database response for FUNCTION LOAD: ({})",
+                err
+            );
+            Err(MediatorError::DatabaseError(
+                "Initialization".into(),
+                format!("Loading LUA scripts, received database error: {}", err),
+            ))
+        }
     }
 }